+++ src/plugins/ncr/actions/receiveNodes.js
@@ -1,0 +1,6 @@
+import { actionTypes } from '../constants';
+
+export default (nodes) => ({
+  type: actionTypes.NODES_RECEIVED,
+  nodes,
+});