+++ src/plugins/raven/components/active-edits-table/delegate.js
@@ -1,0 +1,5 @@
+import updateCollaborations from '@triniti/cms/plugins/raven/actions/updateCollaborations';
+
+export default (dispatch) => ({
+  handleUpdateCollaborations: (accessToken) => dispatch(updateCollaborations(accessToken)),
+});